requirements.txt
selenium
robotframework
pyvirtualdisplay
pyscreenshot
Pillow
robotframework-selenium2library